refactored Inventory with cake pattern
