--- src/main/scala/com/bundlepricing/repos/ItemRepo.scala
@@ -1,10 +1,0 @@
-package com.bundlepricing.repos
-
-import com.bundlepricing.domains.Item
-
-class ItemRepo extends RepoMetaData with InMemoryRepository {
-  type Id = String
-  type Entity = Item
-  
-  def id(entity: Item): Id = entity.name
-}
