+++ src/main/scala/com/bundlepricing/repos/ItemRepoComponent.scala
@@ -1,0 +1,14 @@
+package com.bundlepricing.repos
+
+import com.bundlepricing.domains.Item
+
+trait ItemRepoComponent {
+  def itemRepo: ItemRepo
+
+  trait ItemRepo extends RepoMetaData with Repository {
+    type Id = String
+    type Entity = Item
+    
+    def id(entity: Item): Id = entity.name
+  }
+}
